cleaned and updated docs
