+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@unstoppabledomains/resolution",
-  "version": "1.8.0",
+  "version": "1.8.1",
   "description": "Domain Resolution for blockchain domains",
   "main": "./build/index.js",
   "directories": {
