+++ src/Cns.ts
@@ -1,9 +1,6 @@
 import { EthereumNamingService } from './EthereumNamingService';
 import {
-  NamingServiceName,
   ReaderMap,
-  ResolutionResponse,
-  SourceDefinition,
   isNullAddress,
   NullAddress,
 } from './types';
@@ -17,6 +14,7 @@ import Contract from './utils/contract';
 import standardKeys from './utils/standardKeys';
 import { isLegacyResolver } from './utils';
 import { isValidTwitterSignature } from './utils/verifyTwitterSig';
+import { SourceDefinition, NamingServiceName, ResolutionResponse } from './publicTypes';
 
 const ReaderMap: ReaderMap = {
   1: '0x7ea9ee21077f84339eda9c80048ec6db678642b1',
