+++ src/EthereumNamingService.ts
@@ -4,13 +4,12 @@ import ResolutionError, { ResolutionErrorCode } from './errors/resolutionError';
 import {
   BlockhanNetworkUrlMap,
   isNullAddress,
-  NamingServiceName,
   NetworkIdMap,
-  SourceDefinition,
   nodeHash,
 } from './types';
 import { invert } from './utils';
 import Contract from './utils/contract';
+import { NamingServiceName, SourceDefinition } from './publicTypes';
 
 /** @internal */
 export abstract class EthereumNamingService extends NamingService {
