+++ src/FetchProvider.ts
@@ -1,7 +1,8 @@
 import { FetchError } from 'node-fetch';
 import BaseConnection from './BaseConnection';
-import { Provider, RequestArguments, ResolutionMethod } from './types';
+import { RequestArguments } from './types';
 import ResolutionError, { ResolutionErrorCode } from './errors/resolutionError';
+import { Provider, ResolutionMethod } from './publicTypes';
 
 /** @internal */
 export default class FetchProvider extends BaseConnection implements Provider {
