+++ src/UdApi.test.ts
@@ -5,7 +5,7 @@ import {
   DefaultUrl,
   mockAPICalls,
   mockAsyncMethod,
-  expectSpyToBeCalled,, CryptoDomainWithTwitterVerification
+  expectSpyToBeCalled, CryptoDomainWithTwitterVerification
 } from './tests/helpers';
 
 beforeEach(() => {
