+++ src/Zns.ts
@@ -7,35 +7,36 @@ import {
 import { invert, set } from './utils';
 import {
   Dictionary,
-  ResolutionResponse,
-  SourceDefinition,
-  UnclaimedDomainResponse,
   ZnsResolution,
-  NamingServiceName,
   isNullAddress,
   nodeHash,
 } from './types';
-import { ResolutionError, ResolutionErrorCode } from './index';
+import { NamingServiceName, ResolutionError, ResolutionErrorCode, ResolutionResponse, SourceDefinition, UnclaimedDomainResponse } from './index';
 import NamingService from './NamingService';
 
+/** @internal */
 const DefaultSource = 'https://api.zilliqa.com';
 
+/** @internal */
 const NetworkIdMap = {
   mainnet: 1,
   testnet: 333,
   localnet: 111,
 };
 
+/** @internal */
 const RegistryMap = {
   1: 'zil1jcgu2wlx6xejqk9jw3aaankw6lsjzeunx2j0jz',
 };
 
+/** @internal */
 const UrlMap = {
   1: 'https://api.zilliqa.com',
   333: 'https://dev-api.zilliqa.com',
   111: 'http://localhost:4201',
 };
 
+/** @internal */
 const UrlNetworkMap = (url: string) => invert(UrlMap)[url];
 
 /** @internal */
