+++ src/ens/contract/resolver.ts
@@ -1,7 +1,8 @@
 import { ResolutionError, ResolutionErrorCode } from '../../errors/resolutionError';
 import { default as newResolver } from './newResolver';
 import { default as oldResolver } from './oldResolver';
-import { EthCoinIndex, NamingServiceName } from '../../types';
+import { EthCoinIndex } from '../../types';
+import { NamingServiceName } from '../../publicTypes';
 
 export const OldResolverAddresses = [
   '0x5ffc014343cd971b7eb70732021e26c35b744cc4',
