+++ src/utils/Eip1993Factories.ts
@@ -1,13 +1,11 @@
 import {
   EthersProvider,
   JsonRpcResponse,
-  Provider,
   RequestArguments,
-  Web3Version0Provider,
-  Web3Version1Provider,
 } from '../types';
 import { ConfigurationError, ConfigurationErrorCode } from '../errors/configurationError';
 import ResolutionError, { ResolutionErrorCode } from '../errors/resolutionError';
+import { Web3Version0Provider, Provider, Web3Version1Provider } from '../publicTypes';
 
 export const Eip1993Factories = {
   fromWeb3Version0Provider,
