+++ src/utils/contract.ts
@@ -1,6 +1,6 @@
 import { Interface, JsonFragment } from '@ethersproject/abi';
-import { Provider, RequestArguments, isNullAddress, EventData } from '../types';
-import FetchProvider from '../FetchProvider';
+import { RequestArguments, EventData } from '../types';
+import { Provider } from '../publicTypes';
 
 /** @internal */
 export default class Contract {
