+++ src/zns/utils.ts
@@ -2,13 +2,18 @@
  * All functionality below came from here https://github.com/Zilliqa/Zilliqa-JavaScript-Library/tree/dev/packages/zilliqa-js-crypto/src
  */
 
+
 import hashjs from 'hash.js';
 import BN from 'bn.js';
 
+/** @internal */
 const CHARSET = 'qpzry9x8gf2tvdw0s3jn54khce6mua7l';
+/** @internal */
 const GENERATOR = [0x3b6a57b2, 0x26508e6d, 0x1ea119fa, 0x3d4233dd, 0x2a1462b3];
 // HRP is the human-readable part of zilliqa bech32 addresses
+/** @internal */
 const HRP = 'zil';
+/** @internal */
 const tHRP = 'tzil';
 
 /** @internal */
@@ -16,6 +21,7 @@ function isByteString(str: string, len: number) {
   return !!str.replace('0x', '').match(`^[0-9a-fA-F]{${len}}$`);
 }
 
+/** @internal */
 function isAddress(address: string) {
   return isByteString(address, 40);
 }
@@ -28,7 +34,7 @@ function isAddress(address: string) {
  * For example, converts byte buffers to buffers of maximum 5 bit numbers,
  * padding those numbers as necessary. Necessary for encoding Ethereum-style
  * addresses as bech32 ones.
- *
+ * @internal 
  * @param {Buffer} data
  * @param {number} fromWidth
  * @param {number} toWidth
@@ -70,6 +76,7 @@ function convertBits(
   return Buffer.from(ret);
 }
 
+/** @internal */
 function hrpExpand(hrp: string): Buffer {
   const ret: any[] = [];
   let p;
@@ -83,6 +90,7 @@ function hrpExpand(hrp: string): Buffer {
   return Buffer.from(ret);
 }
 
+/** @internal */
 function polymod(values: Buffer): number {
   let chk = 1;
   // tslint:disable-next-line
@@ -98,6 +106,7 @@ function polymod(values: Buffer): number {
   return chk;
 }
 
+/** @internal */
 function createChecksum(hrp: string, data: Buffer) {
   const values = Buffer.concat([
     Buffer.from(hrpExpand(hrp)),
@@ -113,10 +122,12 @@ function createChecksum(hrp: string, data: Buffer) {
   return Buffer.from(ret);
 }
 
+/** @internal */
 function verifyChecksum(hrp: string, data: Buffer) {
   return polymod(Buffer.concat([hrpExpand(hrp), data])) === 1;
 }
 
+/** @internal */
 function encode(hrp: string, data: Buffer) {
   const combined = Buffer.concat([data, createChecksum(hrp, data)]);
   let ret = hrp + '1';
@@ -127,6 +138,7 @@ function encode(hrp: string, data: Buffer) {
   return ret;
 }
 
+/** @internal */
 function decode(bechString: string) {
   let p;
   let hasLower = false;
@@ -171,7 +183,7 @@ function decode(bechString: string) {
  * toChecksumAddress
  *
  * takes hex-encoded string and returns the corresponding address
- *
+ * @internal
  * @param {string} address
  * @returns {string}
  */
@@ -209,7 +221,7 @@ export const toChecksumAddress = (address: string): string => {
  *
  * The expected format is zil1<address><checksum> where address and checksum
  * are the result of bech32 encoding a Buffer containing the address bytes.
- *
+ * @internal 
  * @param {string} 20 byte canonical address
  * @returns {string} 38 char bech32 encoded zilliqa address
  */
@@ -236,7 +248,7 @@ export function toBech32Address(
 
 /**
  * fromBech32Address
- *
+ * @internal
  * @param {string} address - a valid Zilliqa bech32 address
  * @returns {string} a canonical 20-byte Ethereum-style address
  */
