+++ tsconfig.json
@@ -19,6 +19,12 @@
     "mode": "file",
     "module": "commonjs",
     "excludePrivate": true,
-    "excludeNotExported": true
+    "excludeNotExported": true,
+    "inputFiles": [
+      "./src/Resolution.ts",
+      "./src/publicTypes.ts",
+      "./src/errors/configurationError.ts",
+      "./src/errors/resolutionError.ts",
+    ]
   }
 }
